requirements.txt
pyaudio
SpeechRecognition
geocoder
twilio
Flask
python-dotenv